requirements.txt
faster-whisper==1.0.1
Flask==2.2.2
Werkzeug==2.2.2